Replace all singles with single-precision floats
At the time of generation I did not realize they were f32.

+++ terraria-protocol/src/packets/play_music_item.rs
@@ -7,7 +7,7 @@ use crate::SliceCursor;
 #[derive(Debug)]
 pub struct PlayMusicItem {
     pub player_id: u8,
-    pub note: i32, /* single */
+    pub note: f32,
 }
 
 impl PacketBody for PlayMusicItem {
